TST: move TestNonzero to numpy_tests/core/test_numeric.py
